Change ID types from i32 to u64 in PostgreSQL teams implementations

+++ src/postgres/teams/context.rs
@@ -18,16 +18,16 @@ impl PostgresUserTeamContextRepository {
 
 #[derive(FromRow)]
 struct ContextRecord {
-    user_id: i32,
-    current_team_id: i32,
+    user_id: i64,
+    current_team_id: i64,
     updated_at: DateTime<Utc>,
 }
 
 impl From<ContextRecord> for UserTeamContext {
     fn from(row: ContextRecord) -> Self {
         UserTeamContext {
-            user_id: row.user_id,
-            current_team_id: row.current_team_id,
+            user_id: row.user_id as u64,
+            current_team_id: row.current_team_id as u64,
             updated_at: row.updated_at,
         }
     }
@@ -36,11 +36,11 @@ impl From<ContextRecord> for UserTeamContext {
 #[async_trait]
 impl UserTeamContextRepository for PostgresUserTeamContextRepository {
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn get_context(&self, user_id: i32) -> Result<Option<UserTeamContext>, AuthError> {
+    async fn get_context(&self, user_id: u64) -> Result<Option<UserTeamContext>, AuthError> {
         let row: Option<ContextRecord> = sqlx::query_as(
             "SELECT user_id, current_team_id, updated_at FROM user_team_contexts WHERE user_id = $1",
         )
-        .bind(user_id)
+        .bind(user_id as i64)
         .fetch_optional(&self.pool)
         .await
         .map_err(|e| {
@@ -54,8 +54,8 @@ impl UserTeamContextRepository for PostgresUserTeamContextRepository {
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
     async fn set_current_team(
         &self,
-        user_id: i32,
-        team_id: i32,
+        user_id: u64,
+        team_id: u64,
     ) -> Result<UserTeamContext, AuthError> {
         let row: ContextRecord = sqlx::query_as(
             r"
@@ -66,8 +66,8 @@ impl UserTeamContextRepository for PostgresUserTeamContextRepository {
             RETURNING user_id, current_team_id, updated_at
             ",
         )
-        .bind(user_id)
-        .bind(team_id)
+        .bind(user_id as i64)
+        .bind(team_id as i64)
         .fetch_one(&self.pool)
         .await
         .map_err(|e| {
@@ -79,9 +79,9 @@ impl UserTeamContextRepository for PostgresUserTeamContextRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn clear_context(&self, user_id: i32) -> Result<(), AuthError> {
+    async fn clear_context(&self, user_id: u64) -> Result<(), AuthError> {
         sqlx::query("DELETE FROM user_team_contexts WHERE user_id = $1")
-            .bind(user_id)
+            .bind(user_id as i64)
             .execute(&self.pool)
             .await
             .map_err(|e| {
