+++ src/postgres/teams/permission.rs
@@ -43,14 +43,14 @@ where
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
     async fn get_permissions(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: u64,
+        user_id: u64,
     ) -> Result<PermissionSet<R, A>, AuthError> {
         let row: Option<PermissionRecord> = sqlx::query_as(
             "SELECT permissions FROM team_member_permissions WHERE team_id = $1 AND user_id = $2",
         )
-        .bind(team_id)
-        .bind(user_id)
+        .bind(team_id as i64)
+        .bind(user_id as i64)
         .fetch_optional(&self.pool)
         .await
         .map_err(|e| {
@@ -73,8 +73,8 @@ where
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self, permissions), err))]
     async fn set_permissions(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: u64,
+        user_id: u64,
         permissions: &PermissionSet<R, A>,
     ) -> Result<(), AuthError> {
         let json_str = permissions.to_json();
@@ -91,8 +91,8 @@ where
             DO UPDATE SET permissions = $3, updated_at = NOW()
             ",
         )
-        .bind(team_id)
-        .bind(user_id)
+        .bind(team_id as i64)
+        .bind(user_id as i64)
         .bind(json_value)
         .execute(&self.pool)
         .await
@@ -110,8 +110,8 @@ where
     )]
     async fn grant_permission(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: u64,
+        user_id: u64,
         resource: R,
         action: A,
     ) -> Result<(), AuthError> {
@@ -126,8 +126,8 @@ where
     )]
     async fn revoke_permission(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: u64,
+        user_id: u64,
         resource: &R,
         action: &A,
     ) -> Result<(), AuthError> {
@@ -142,8 +142,8 @@ where
     )]
     async fn has_permission(
         &self,
-        team_id: i32,
-        user_id: i32,
+        team_id: u64,
+        user_id: u64,
         resource: &R,
         action: &A,
     ) -> Result<bool, AuthError> {
